List operator-backed services in the sidebar
Closes #3494
